+++ src/constructs/iam/fastly-logs-iam.test.ts
@@ -1,6 +1,6 @@
 import "@aws-cdk/assert/jest";
 import { SynthUtils } from "@aws-cdk/assert";
-import { simpleGuStackForTesting } from "../utils/test";
+import { simpleGuStackForTesting } from "../../utils/test";
 import { GuFastlyLogsIam } from "./fastly-logs-iam";
 
 describe("The GuFastlyLogsIam pattern", () => {
