+++ src/constructs/iam/index.ts
@@ -1,2 +1,3 @@
 export * from "./policies";
 export * from "./roles";
+export * from "./fastly-logs-iam";
