+++ src/patterns/index.ts
@@ -3,4 +3,3 @@ export * from "./sns-lambda";
 export * from "./kinesis-lambda";
 export * from "./ec2-app";
 export * from "./scheduled-ecs-task";
-export * from "./fastly-logs-iam";
